backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic-settings==2.10.1
openai==1.107.1
whisper==1.1.10
pywinrm==0.5.0
requests==2.32.5
